[REFACTOR] Make getCrimstoneStage pure + break the Crimstone import cycle
- Thread the component clock (useNow `now`) into getCrimstoneStage instead of
  calling Date.now() inside it, keeping the stage calc pure in the render path.
- Move getCrimstoneStage into its own leaf module (crimstone/getCrimstoneStage.ts)
  to break the import cycle: Crimstone.tsx imports its Recovered/Depleting/Depleted
  children, and each imported getCrimstoneStage back from Crimstone.tsx.
